+++ lib/dchs_flutter_beacon_platform_interface.dart
@@ -43,6 +43,10 @@ abstract class DchsFlutterBeaconPlatform extends PlatformInterface {
   Future<bool> get openApplicationSettings;
   Future<bool> setScanPeriod(int scanPeriod);
   Future<bool> setBetweenScanPeriod(int scanPeriod);
+  Future<bool> setBackgroundScanPeriod(int scanPeriod);
+  Future<bool> setBackgroundBetweenScanPeriod(int scanPeriod);
+  Future<bool> setUseTrackingCache(bool enable);
+  Future<bool> setMaxTrackingAge(int age);
   Future<bool> get close;
   Stream<RangingResult> ranging(List<Region> regions);
   Stream<MonitoringResult> monitoring(List<Region> regions);